refactor: useJoinGroupMutation과 fetchJoinGroup 분리
