+++ src/app/(afterLogin)/home/hooks/useJoinGroupMutation.ts
@@ -3,24 +3,14 @@ import { groupModalApiData } from '@/app/_types/Api';
 // import { groupModalData } from '@/app/_types/Api';
 import { useMutation, useQueryClient } from  '@tanstack/react-query';
 import { fetchWithAuth } from '../../_lib/FetchWithAuth';
-export async function fetchjoinGroup(groupId:number) {
-const response = await fetchWithAuth(
-  `/groups/${groupId}/participation`,
-  {
-    method:'POST',
-    credentials: 'include',
-    cache: "no-cache",
-  },
-);
-return await response;
-}
+import { fetchJoinGroup } from '../_lib/fetchJoinGroup';
 export function useJoinGroupMutation(groupId: number) {
   
   const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn:  async () => {
-      return await fetchjoinGroup(groupId);
+      return await fetchJoinGroup(groupId);
     },
     onError: (err) => {
       alert(`오류가 발생했습니다: ${err.message}`);
